+++ src/test/java/com/vectorsearch/faiss/TwoIVFFlat.java
@@ -32,7 +32,7 @@ public class TwoIVFFlat extends FaissTestCase {
         index.train(inputRowCount, inputVectors.cast());
         index.add(inputRowCount, inputVectors.cast());
         final boolean isTrained = index.getIs_trained();
-        final int nTotal = index.getNtotal();
+        final long nTotal = index.getNtotal();
         final String msg = "isTrained = " + isTrained + ", nTotal = " + nTotal;
         LOGGER.info(msg);
     }
@@ -43,7 +43,7 @@ public class TwoIVFFlat extends FaissTestCase {
         final floatArray distances = new floatArray(queryRowCount * rn);
         final longArray indices = new longArray(queryRowCount * rn);
         index.setNprobe(10);
-        index.search(queryRowCount, queryVectors.cast(), rn, distances.cast(), indices.cast());
+        index.search((long) queryRowCount, queryVectors.cast(), (long) rn, distances.cast(), indices.cast());
         LOGGER.info(show(distances, 5, rn));
         LOGGER.info(show(indices, 5, rn));
     }
